Fixed build for Lua 5.2 ("make test" works again).
1. updated srlua for 5.2 (but this may have broken 5.1?)
2. updated luatest.lua to 5.1/5.2 compatible version.
3. fixed Makefile a bit for OS X.

+++ runtime/parse.c
@@ -44,13 +44,13 @@ void dump_stack(struct gzl_parse_state *s, FILE *output)
 
             case GZL_FRAME_TYPE_GLA: {
                 struct gzl_gla_frame *gla_frame = &frame->f.gla_frame;
-                fprintf(output, "GLA: #%d, ", gla_frame->gla - g->glas);
+                fprintf(output, "GLA: #%d, ", (int)(gla_frame->gla - g->glas));
                 break;
             }
 
             case GZL_FRAME_TYPE_INTFA: {
                 struct gzl_intfa_frame *intfa_frame = &frame->f.intfa_frame;
-                fprintf(output, "IntFA: #%d, ", intfa_frame->intfa - g->intfas);
+                fprintf(output, "IntFA: #%d, ", (int)(intfa_frame->intfa - g->intfas));
                 break;
             }
         }
